+++ SRP/ControlRoom/Modules/Reports/ProgramByBranch.aspx.cs
@@ -7,6 +7,7 @@ using System.Collections.Generic;
 using System.Data;
 using System.Data.SqlClient;
 using System.Linq;
+using System.Text;
 using System.Web;
 using System.Web.UI;
 using System.Web.UI.WebControls;
@@ -28,13 +29,18 @@ namespace GRA.SRP.ControlRoom.Modules.Reports
         {
             MasterPage.RequiredPermission = 4200;
             MasterPage.IsSecure = true;
-            AlertPanel.Visible = false;
+            MasterPage.PageTitle = "Program By Branch Report";
+            AlertPanel.Visible = !string.IsNullOrEmpty(AlertMessage.Text);
 
             if (!IsPostBack)
             {
                 SetPageRibbon(StandardModuleRibbons.ReportsRibbon());
             }
+        }
 
+        protected void DownloadReport_Click(object sender, EventArgs e)
+        {
+            AlertMessage.Text = string.Empty;
             string reportName = ReportName;
             string reportSproc = ReportStoredProcedure;
 
@@ -46,20 +52,35 @@ namespace GRA.SRP.ControlRoom.Modules.Reports
                 tenantName = tenant.LandingName;
             }
 
-            if (!string.IsNullOrEmpty(Request.QueryString["prelogging"]))
+            var parameters = new List<SqlParameter>();
+            parameters.Add(new SqlParameter("@TenID", CRTenantID == null ? -1 : CRTenantID));
+
+            var dataAsOf = new StringBuilder();
+            dataAsOf.AppendFormat("This report was run on {0}", DateTime.Now);
+
+            if (PreloggingOnly.Checked)
             {
                 reportName = AltReportName;
                 reportSproc = AltReportStoredProcedure;
+                dataAsOf.Append(" and contains data up to each program's Logging Start Date");
+            }
+            else if (!string.IsNullOrWhiteSpace(EndDate.Text))
+            {
+                DateTime endDate = DateTime.MinValue;
+                if (DateTime.TryParse(EndDate.Text, out endDate))
+                {
+                    parameters.Add(new SqlParameter("@EndDate", endDate));
+                    reportName = string.Format("{0}Upto{1:yyyyMMddhhmmtt}",
+                        reportName,
+                        endDate);
+                    dataAsOf.AppendFormat(" and contains data up to {0}", endDate);
+                }
             }
 
             var ds = SqlHelper.ExecuteDataset(conn,
                 CommandType.StoredProcedure,
                 reportSproc,
-                new SqlParameter[]
-                    {
-                        new SqlParameter("@TenID", CRTenantID == null ? -1 : CRTenantID)
-                    }
-            );
+                parameters.ToArray());
 
             // fix table names
             if (ds.Tables != null && ds.Tables.Count >= 1 && ds.Tables[0].Rows.Count != 0)
@@ -98,17 +119,20 @@ namespace GRA.SRP.ControlRoom.Modules.Reports
                         completionRate = string.Format("{0:N2}%",
                             Convert.ToDouble(achieverTotal) * 100 / Convert.ToDouble(signupTotal));
                     }
-                    var completionRow = table.NewRow();
-                    completionRow[0] = string.Format("Completion rate: {0}", completionRate);
-                    table.Rows.Add(completionRow);
+                    var additionalRow = table.NewRow();
+                    additionalRow[0] = string.Format("Completion rate: {0}", completionRate);
+                    table.Rows.Add(additionalRow);
+
+                    additionalRow = table.NewRow();
+                    additionalRow[0] = dataAsOf.ToString();
+                    table.Rows.Add(additionalRow);
 
                     count++;
                 }
 
                 CreateExcelFile.CreateExcelDocument(
                     ds,
-                    string.Format("{0}-{1}.xlsx",
-                        DateTime.Now.ToString("yyyyMMdd"),
+                    string.Format("{0}.xlsx",
                         reportName.Replace(" ", string.Empty)),
                     Response);
             }