+++ SRP/ControlRoom/Modules/Reports/ProgramByBranch.aspx.designer.cs
@@ -13,6 +13,24 @@ namespace GRA.SRP.ControlRoom.Modules.Reports {
     public partial class ProgramByBranch {
         
         /// <summary>
+        /// EndDate control.
+        /// </summary>
+        /// <remarks>
+        /// Auto-generated field.
+        /// To modify move field declaration from designer file to code-behind file.
+        /// </remarks>
+        protected global::System.Web.UI.WebControls.TextBox EndDate;
+        
+        /// <summary>
+        /// PreloggingOnly control.
+        /// </summary>
+        /// <remarks>
+        /// Auto-generated field.
+        /// To modify move field declaration from designer file to code-behind file.
+        /// </remarks>
+        protected global::System.Web.UI.WebControls.CheckBox PreloggingOnly;
+        
+        /// <summary>
         /// AlertPanel control.
         /// </summary>
         /// <remarks>